Print types reasonably nicely.
Implementation assumes that there will be a line that prints the type
of the current value. It then assumes that if a type annotation takes
more than one line then each line after the first will start with
spaces and the next line will not.

Should probably use the built-in parser/printer in the future.